py/requirements.txt
argparse>=1.4.0
py-color-log>=1.0.4
mutagen>=1.46.0